requirements.txt
streamlit
torch
torchvision
matplotlib
seaborn
scikit-learn
numpy
Pillow